Move some methods from UnsafeHelper to SqlTypesUnsafeHelper
* UnsafeHelper now only handles binary things
