+++ common/unsafe/src/main/java/org/apache/spark/unsafe/UnsafeHelper.java
@@ -17,9 +17,6 @@
 
 package org.apache.spark.unsafe;
 
-import org.apache.spark.unsafe.types.CalendarInterval;
-import org.apache.spark.unsafe.types.UTF8String;
-
 import java.nio.ByteBuffer;
 
 public final class UnsafeHelper {
@@ -31,22 +28,6 @@ public final class UnsafeHelper {
     final byte[] bytes = new byte[size];
     Platform.copyMemory(baseObject, baseOffset + offset, bytes, Platform.BYTE_ARRAY_OFFSET, size);
     return bytes;
-  }
-
-  public static UTF8String getUTF8String(long offsetAndSize, Object baseObject, long baseOffset) {
-    final int offset = getOffsetFromOffsetAndSize(offsetAndSize);
-    final int size = getSizeFromOffsetAndSize(offsetAndSize);
-    return UTF8String.fromAddress(baseObject, baseOffset + offset, size);
-  }
-
-  public static CalendarInterval getInterval(
-      long offsetAndSize,
-      Object baseObject,
-      long baseOffset) {
-    final int offset = getOffsetFromOffsetAndSize(offsetAndSize);
-    final int months = (int) Platform.getLong(baseObject, baseOffset + offset);
-    final long microseconds = Platform.getLong(baseObject, baseOffset + offset + 8);
-    return new CalendarInterval(months, microseconds);
   }
 
   public static void writeTo(
