+++ sql/catalyst/src/main/java/org/apache/spark/sql/catalyst/expressions/UnsafeRow.java
@@ -405,7 +405,7 @@ public final class UnsafeRow extends InternalRow implements Externalizable, Kryo
   @Override
   public UTF8String getUTF8String(int ordinal) {
     if (isNullAt(ordinal)) return null;
-    return UnsafeHelper.getUTF8String(getLong(ordinal), baseObject, baseOffset);
+    return SqlTypesUnsafeHelper.getUTF8String(getLong(ordinal), baseObject, baseOffset);
   }
 
   @Override
@@ -422,7 +422,7 @@ public final class UnsafeRow extends InternalRow implements Externalizable, Kryo
     if (isNullAt(ordinal)) {
       return null;
     } else {
-      return UnsafeHelper.getInterval(getLong(ordinal), baseObject, baseOffset);
+      return SqlTypesUnsafeHelper.getInterval(getLong(ordinal), baseObject, baseOffset);
     }
   }
